% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_score_max}
\alias{add_score_max}
\title{Format a diagnostic table}
\usage{
add_score_max(diagnostic.table, k_range)
}
\arguments{
\item{diagnostic.table}{a diagnostic table, each column with a different diagnostics.}

\item{k_range}{the range of k values on which to restrict the max}
}
\value{
The formatted diagnostic table
}
\description{
Format a diagnostic table, and add the max score.
}
\keyword{internal}
