% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calDIST}
\alias{calDIST}
\title{Distances of a vector to rows of a matrix}
\usage{
calDIST(y, x)
}
\arguments{
\item{y}{vector of numeric values}

\item{x}{matrix of row vectors}
}
\value{
vector of distances between y and each row of x
}
\description{
Distances of a vector to rows of a matrix
}
\keyword{internal}
