% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_param_transform}
\alias{check_param_transform}
\title{find transform functions}
\usage{
check_param_transform(
  param,
  param_transform,
  param_lower_bound,
  param_upper_bound
)
}
\arguments{
\item{param}{the matrix of parameters matching the sumstat.}

\item{param_transform}{a named vector of function, that to each parameter attributes a transformation function.
Each function must be one of "none" (no transformation), "log" or "logit".
If "logit", upper and lower values are taken in \code{param_lower_bound} \code{param_upper_bound} arguments.
If the vector is of length 1, it will be recycled to all parameters.
Default to "none": no transformation on any parameter.}

\item{param_lower_bound}{a named vector of lower bounds for each parameter in \code{param}.
If the vector is of length 1, it will be recycled to all parameters.
Default to \code{-Inf}: no lower bound.}

\item{param_upper_bound}{a named vector of lower bounds for each parameter in \code{param}.
If the vector is of length 1, it will be recycled to all parameters.
Default to \code{Inf}: no upper bound.}
}
\value{
a list, with:
\itemize{
\item \code{transform} a vector of transform for each parameter
\item \code{back_transform} a vector of back transform for each parameter.
}
}
\description{
find transform functions
}
\keyword{internal}
