% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_replica}
\alias{check_replica}
\title{Check replicate}
\usage{
check_replica(target, target.replica)
}
\arguments{
\item{target}{the target matrix}

\item{target.replica}{the sumstat matrix}
}
\value{
nothing
}
\description{
Check a matrix and its replicate.
}
\keyword{internal}
