% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sumstat}
\alias{check_sumstat}
\title{Check sumstat}
\usage{
check_sumstat(target, sumstat)
}
\arguments{
\item{target}{the target matrix}

\item{sumstat}{the sumstat matrix}
}
\value{
The formatted sumstat matrix.
}
\description{
Format the sumstat matrix.
}
\keyword{internal}
