% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_target}
\alias{check_target}
\title{Check target}
\usage{
check_target(target)
}
\arguments{
\item{target}{the target matrix}
}
\value{
The formatted target matrix.
}
\description{
Format the target matrix.
}
\keyword{internal}
