% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_power}
\alias{compute_power}
\title{Compute power}
\usage{
compute_power(gfit_obj, level = 0.05)
}
\arguments{
\item{gfit_obj}{an object from the \code{\link{gfit}} or \code{\link{postgfit}} functions.}

\item{level}{the level of the test. Default to 0.05.}
}
\value{
The same object, with an added power slot.
}
\description{
Computes the power of a test.
}
