% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_power_sd}
\alias{compute_power_sd}
\title{Compute power}
\usage{
compute_power_sd(gfit_obj)
}
\arguments{
\item{gfit_obj}{an object from the \code{\link{gfit}} or \code{\link{postgfit}} functions.}
}
\value{
The same object, with an added pow_sd slot.
}
\description{
Computes the approximate standard deviation of the power.
This estimate ignores the variability of the pvalues themselves, and is probably too small.
}
