% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_pval_sd}
\alias{compute_pval_sd}
\title{Compute power}
\usage{
compute_pval_sd(gfit_obj)
}
\arguments{
\item{gfit_obj}{an object from the \code{\link{gfit}} or \code{\link{postgfit}} functions.}
}
\value{
The same object, with an added pval_sd slot.
}
\description{
Computes the approximate standard deviation of pvalues
}
