% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_diagnostic}
\alias{format_diagnostic}
\title{Format a diagnostic table}
\usage{
format_diagnostic(diagnostic.table, scores)
}
\arguments{
\item{diagnostic.table}{a diagnostic table, each column with a different diagnostics.}

\item{scores}{the different scores per diagnostic.}
}
\value{
The formatted diagnostic table
}
\description{
Format a diagnostic table, and add the max score.
}
\keyword{internal}
