% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqgfit.R
\name{freqgfit}
\alias{freqgfit}
\title{Pre-Computed Holdout Frequentist Goodness of Fit Test}
\usage{
freqgfit(
  target,
  target.replica,
  sumstat,
  sumstat.replica,
  score = c("lof", "kNN"),
  k = c(2, 3, 5, 8, 14, 20),
  k_range = range(k),
  eps = 0.01,
  split = 0.5,
  norm = sd,
  ncores = 1
)
}
\arguments{
\item{target}{a matrix of observations, with dimension Nobs x Nstats.}

\item{target.replica}{a matrix of observations, with dimension Nobs x Nstats.
Each line i of the replicate target should come from the same model
as the one used for the same line i of the reference table.}

\item{sumstat}{a matrix of summary statistics from the training set,
with dimension Ntrain x Nstats.}

\item{sumstat.replica}{a matrix of replicate summary statistics from the training set,
with dimension Ntrain x Nstats.
Each line i of the replicate summary statistic should have been simulated
with the same parameters as the one used for the same line i of the reference table.}

\item{score}{the score to use for calling outliers. Can be "lof", "kNN", or both.}

\item{k}{The kth-distance to be used to calculate LOFs of kNNs.
k can be a vector which contains multiple k values based on which the score needs to be calculated.}

\item{k_range}{A vector of min and max values for k when using the "LOF max" score (see Details).}

\item{eps}{the proportion of data points used to approximate the posterior distribution
of the reference table around each observation point. Default to 0.01.}

\item{split}{proportion of the posterior to be used for calibration. Default to 0.5.}

\item{norm}{the normalization function. Default to \code{\link{sd}}.}

\item{ncores}{number of cores for parallel computations.}
}
\value{
For each score, a list with:
\itemize{
\item \code{pval} a matrix of p-values for the test data (rows), for each possible value of k (if \code{nboot=1}),
or an array of such matrices (if \code{nboot>1}).
\item \code{n.ref} the number of particles in the reference dataset (number of rows of \code{sumstat}).
\item \code{n.calib} the number of particles used for calibration.
\item \code{eps} the localization fraction.
\item \code{split} the fraction of data used for calibration.
\item \code{k_range} the min and max values for k when using the "LOF max" score.
}
}
\description{
This function is similar to \code{\link[=hpgfit]{hpgfit()}}, but it only works for the
rejection method, and for datasets with pre-computed replicates.

It is likely to be DEPRECATED in stable versions of the package.

This function performs a post-inference holdout Goodness of Fit (GoF) test,
that checks whether we can reject the hypothesis that the \code{target}
is from the same model as the particles from the \code{sumstat} matrix.

It works by first approximating the posterior distribution using a
localization parameter \code{eps} the rejection
method (see documentation from \code{abc::abc}).
It then draws a fraction \code{split} of the reference table \code{sumstat}
for calibration, and then compares the score of the
\code{target} with the distribution of scores of the calibration points.
}
\references{
Le Mailloux G., Bastide, P., Marin, J-M., Estoup, A. (2024+),
Goodness of Fit for Bayesian Generative Models with Applications in Population Genetics.

Markus M. Breunig, Hans-Peter Kriegel, Raymond T. Ng, and Jörg Sander. 2000.
LOF: identifying density-based local outliers. SIGMOD Rec. 29, 2 (June 2000), 93–104.
https://doi.org/10.1145/335191.335388
}
\seealso{
\code{\link[=hpgfit]{hpgfit()}} for post-inference holdout GoF test.
}
