% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaussian_laplace}
\alias{gaussian_laplace}
\title{Gaussian Laplace Toy dataset}
\format{
Two lists \code{dataset_gaussian} and \code{dataset_laplace}, with entries:
\describe{
\item{param}{A matrix with 100,000 rows and 2 columns with the parameters used for the simulations}
\item{sim}{A data frame with the simulated summary statistics (350 columns) for each of the 100,000 rows of the param matrix.}
\item{sim.fun}{The simulation function used.}
}
}
\usage{
gaussian_laplace
}
\description{
A dataset containing two reference tables, simulated using either the
a Gaussian or a Laplace model.
}
\details{
Each dataset contains 1000 lines.
Parameters \code{mu} and \code{sigma} were drawn uniformly between -10 and 10, and 1 and 4, respectively.
For each set of parameter:
\itemize{
\item a Gaussian vector of size 350 was simulated with mean \code{mu} and standard deviation \code{sigma} using base function \code{stats::rnorm}.
\item a Laplace vector of size 350 was simulated with location \code{mu} and scale \code{sigma/sqrt(2)} using function \code{VGAM::rlaplace}.
Then, for each vector, a set of 20 summary statistics was computed, taking the 20 first normalized L-moments using function \code{lmom::samlmu}.
}
}
\keyword{datasets}
