% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_p_val_vec}
\alias{get_p_val_vec}
\title{Get p value from diagnostic}
\usage{
get_p_val_vec(diagnostic.calib, diagnostic.test)
}
\arguments{
\item{diagnostic.calib}{a matrix of calibration scores, size Ncalib x Nscores.}

\item{diagnostic.test}{a vector of test scores, length Nscores.}
}
\value{
A vector of p values of length Nscores.
}
\description{
Computes the empirical p values for a set of scores.
}
\keyword{internal}
