% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfit.R
\name{gfit}
\alias{gfit}
\alias{print.gfit}
\alias{summary.gfit}
\alias{print.gfitsummary}
\title{Prior Goodness of Fit test}
\usage{
gfit(
  target,
  sumstat,
  nb.replicate = 100,
  score = c("lof", "kNN"),
  k = c(2, 3, 5, 8, 14, 20),
  k_range = range(k),
  norm = sd,
  ncores = 1,
  nboot = 1
)

\method{print}{gfit}(x, ...)

\method{summary}{gfit}(object, score = "lof", k = "max", level = 0.95, ...)

\method{print}{gfitsummary}(x, ...)
}
\arguments{
\item{target}{a matrix of observations, with dimension Nobs x Nstats.}

\item{sumstat}{a matrix of summary statistics from the training set,
with dimension Ntrain x Nstats.}

\item{nb.replicate}{number of replicates to draw from the training set for the calibration of the p value.}

\item{score}{the score to use for calling outliers. Can be "lof", "kNN", or both.}

\item{k}{The kth-distance to be used to calculate LOFs of kNNs.
k can be a vector which contains multiple k values based on which the score needs to be calculated.}

\item{k_range}{A vector of min and max values for k when using the "LOF max" score (see Details).}

\item{norm}{the normalization function. Default to \code{\link{sd}}.}

\item{ncores}{number of cores for parallel computations.}

\item{nboot}{number of bootstrap replicates (see Details). Default to 1: no bootstrap replicate.}

\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{an object for which a summary is desired.}

\item{level}{the level for the confidence interval}
}
\value{
An object of class \code{gfit}. For each score, it contains a list with:
\itemize{
\item \code{score.calib} a matrix of LOF scores for the calibration data (rows), for each possible value of k (if \code{nboot=1}).
\item \code{score.test} a matrix of LOF scores for the test data (rows), for each possible value of k (if \code{nboot=1}).
\item \code{pval} a matrix of p-values for the test data (rows), for each possible value of k (if \code{nboot=1}),
or an array of such matrices (if \code{nboot>1}).
\item \code{n.ref} the number of particles in the reference dataset (number of rows of \code{sumstat}).
\item \code{n.calib} the number of particles used for calibration (equal to \code{nb.replicate}).
\item \code{n.boot} the number of bootstrap replicates.
\item \code{k_range} the min and max values for k when using the "LOF max" score.
}
}
\description{
This function performs a prior Goodness of Fit (GoF) test,
that checks whether we can reject the hypothesis that the \code{target}
is from the same model as the particles from the \code{sumstat} matrix.

It works by sub-sampling \code{nb.replicate} points from the
reference table \code{sumstat}, and then by comparing the score of the
\code{target} with the distribution of scores of the calibration points.
}
\details{
The LOF and kNN scores are computations rely on the
function \code{\link[dbscan]{kNN}} from the \code{dbscan} package.

\code{target} can be a matrix, in which case the GoF test is performed independently
for each observation vector (rows of the matrix).

When using the "LOF max" statistics, the maximum LOF score over \code{k_range}
is taken for all calibration and observation points.

When \code{nboot>1}, the test is performed several times, by drawing \code{nboot}
different sets of \code{nb.replicates} calibration points.
Highest Density Intervals are then computed using the
\code{\link[HDInterval]{hdi}} function from package \code{HDInterval}.
}
\examples{
data(gaussian_laplace)
## 1000 simulations from the Laplace dataset
sumstat <- gaussian_laplace$dataset_laplace$sim
## 10 observations from the Gaussian dataset
target <- gaussian_laplace$dataset_gaussian$sim[1:10, ]

## Apply gfit
resgfit <- gfit(target, sumstat)
resgfit

## Apply gfit with bootstrap
resgfit <- gfit(target, sumstat, nboot = 10) # nboot should be larger
resgfit

}
\references{
Le Mailloux G., Bastide, P., Marin, J-M., Estoup, A. (2024+),
Goodness of Fit for Bayesian Generative Models with Applications in Population Genetics.

Markus M. Breunig, Hans-Peter Kriegel, Raymond T. Ng, and Jörg Sander. 2000.
LOF: identifying density-based local outliers. SIGMOD Rec. 29, 2 (June 2000), 93–104.
https://doi.org/10.1145/335191.335388
}
\seealso{
\code{\link[=summary.gfit]{summary.gfit()}} for summarizing the results;
\code{\link[=gof.fit]{gof.fit()}} for prior GoF computation with fixed calibration points;
\code{\link[=hpgfit]{hpgfit()}} for post-inference holdout GoF test.
}
