% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpgfit.R
\name{gof.fit.holdout}
\alias{gof.fit.holdout}
\title{"Frequentist" goodness of fit function}
\usage{
gof.fit.holdout(
  data.test.obs,
  data.test.new,
  data.param,
  data.ref,
  sim.fun,
  method = c("loclinear", "ridge", "rejection"),
  kernel = c("epanechnikov", "rectangular", "gaussian", "triangular", "biweight",
    "cosine"),
  lambda = c(1e-04, 0.001, 0.01),
  trans.fun = function(x) return(x),
  back.trans.fun = function(x) return(x),
  k = c(2, 3, 5, 8, 14, 20),
  k_range = range(k),
  eps = 1/100,
  split = 0.5,
  normvec,
  score = c("lof", "kNN"),
  ncores = 1,
  nboot = 1,
  ...
)
}
\arguments{
\item{data.test.obs}{a normalized matrix of observations, with dimension Nobs x Nstats.}

\item{data.test.new}{a normalized matrix of new (replicated) observations, with dimension Nobs x Nstats.}

\item{data.param}{the matrix of parameters matching the sumstat.}

\item{data.ref}{a normalized matrix of reference summary statistics from the training set, with dimension Ntrain x Nstats.}

\item{sim.fun}{a function that, for a matrix of parameters (size nsim x nparams), returns a simulated matrix of sumstats (size nsim x nsumstats).}

\item{method}{one of "rejection" (default), "loclinear" or "ridge".}

\item{kernel}{for "loclinear" and "ridge", the kernel to be used.}

\item{lambda}{for "ridge", the lambda penalties to be used in the ridge regression. The median of all predictions with various lambdas will be used. Default to \code{lambda = c(0.0001, 0.001, 0.01)}.}

\item{trans.fun}{a transformation function, that takes a matrix of parameters, and return a transformed matrix of parameters.}

\item{back.trans.fun}{a transformation function, that takes a transformed matrix of parameters, and return a matrix of parameters in the original space.}

\item{k}{The kth-distance to be used to calculate LOFs of kNNs.
k can be a vector which contains multiple k values based on which the score needs to be calculated.}

\item{k_range}{A vector of min and max values for k when using the "LOF max" score (see Details).}

\item{eps}{the proportion of data points used to approximate the posterior distribution
of the reference table around each observation point. Default to 0.01.}

\item{split}{proportion of the posterior to be used for calibration. Default to 0.5.}

\item{normvec}{the column normalization values.}

\item{score}{the score to use for calling outliers. Can be "lof", "kNN", or both.}

\item{ncores}{number of cores for parallel computations.}

\item{nboot}{number of bootstrap replicates (see Details). Default to 1: no bootstrap replicate.}

\item{...}{further arguments to be passed to \code{sim.fun}.}
}
\value{
For each score, a list, with
\itemize{
\item \code{pval} a matrix of p-values for the test data (rows), for each possible value of k.
}
}
\description{
Internal function to compute the GoF using the LOF scores.
The data should already be formatted and normalized.
Unless you are an advanced user, you should probably use
\code{\link{hpgfit}}.
}
