% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgfit.R
\name{gof.fit.post}
\alias{gof.fit.post}
\title{"Posterior" goodness of fit function}
\usage{
gof.fit.post(
  data.test,
  data.calib,
  data.ref,
  data.ref.replica,
  k = c(2, 3, 5, 8, 14, 20),
  k_range = range(k),
  eps = 1/100,
  score = c("lof", "kNN"),
  ncores = 1
)
}
\arguments{
\item{data.test}{a normalized matrix of observations, with dimension Nobs x Nstats.}

\item{data.calib}{a normalized matrix of reference summary statistics from the training set, with dimension Ncalib x Nstats.}

\item{data.ref}{a normalized matrix of reference summary statistics from the training set, with dimension Ntrain x Nstats.}

\item{data.ref.replica}{a normalized matrix of reference replicate summary statistics
from the training set, with dimension Ntrain x Nstats.
Each line i of the replicate summary statistic should have been simulated
with the same parameters as the one used for the same line i of the reference table.}

\item{k}{The kth-distance to be used to calculate LOFs of kNNs.
k can be a vector which contains multiple k values based on which the score needs to be calculated.}

\item{k_range}{A vector of min and max values for k when using the "LOF max" score (see Details).}

\item{eps}{the proportion of data points used to approximate the posterior distribution
of the reference table around each observation point. Default to 0.01.}

\item{score}{the score to use for calling outliers. Can be "lof", "kNN", or both.}

\item{ncores}{number of cores for parallel computations.}
}
\value{
For each score, it a list with:
\itemize{
\item \code{score.calib} a matrix of LOF scores for the calibration data (rows), for each possible value of k.
\item \code{score.test} a matrix of LOF scores for the test data (rows), for each possible value of k.
\item \code{pval} a matrix of p-values for the test data (rows), for each possible value of k.
\item \code{n.ref} the number of particles in the reference dataset (number of rows of \code{sumstat}).
\item \code{n.calib} the number of particles used for calibration (equal to \code{nb.replicate}).
\item \code{k_range} the min and max values for k when using the "LOF max" score.
}
}
\description{
Internal function to compute the posterior GoF.
The data should already be formatted and normalized.
Unless you are an advanced user, you should probably use
\code{\link{postgfit}}.
}
