% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof.internal.R
\name{kdist.new}
\alias{kdist.new}
\title{New k-dist}
\usage{
kdist.new(old.kdist, old.belowkdist, dist.new)
}
\arguments{
\item{old.kdist}{k-dist computed without the new points.}

\item{old.belowkdist}{(k-1)-dist computed without the new points.}

\item{dist.new}{distance to the new point}
}
\value{
new k-dist including the new point.
}
\description{
Compute the new k-dist for a point in the reference data set. It uses the old k-distances and the distance to the new point.
}
\keyword{internal}
