% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof.internal.R
\name{lof.internal}
\alias{lof.internal}
\title{LOF one new point}
\usage{
lof.internal(new, set.ref, k_all, Tab.dist = NULL, Tab.id = NULL)
}
\arguments{
\item{new}{The vector that represents the new point}

\item{set.ref}{The ensemble of points in the reference data set  (row by row : \code{set.ref[j,]} represents the j_th point)}

\item{k_all}{specify the values of k}

\item{Tab.dist}{\code{Tab.dist[i,k]} k-distance of particles i}

\item{Tab.id}{\code{Tab.id[i,1:k]} k-neighborhood of particles i}
}
\value{
vector of LOF values for the new point with the respect to the reference data set, for each k in k_all.
}
\description{
Compute the LOF for the new point with respect to the data set of reference for various values of k.
You need to have pre-computed k_dist for points in the reference data set.
}
\keyword{internal}
