% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic}
\alias{logistic}
\title{logistic transform}
\usage{
logistic(x, a = 0, b = 1)
}
\arguments{
\item{x}{vector of values}

\item{a}{lower bound}

\item{b}{upper bound}
}
\value{
transformed values
}
\description{
logistic transform
}
\keyword{internal}
