% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logisticVec}
\alias{logisticVec}
\title{logistic transform}
\usage{
logisticVec(x, a = 0, b = rep(1, length(x)))
}
\arguments{
\item{x}{vector of values}

\item{a}{lower bound}

\item{b}{vector of upper bound}
}
\value{
transformed values
}
\description{
logistic transform
}
\keyword{internal}
