% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof.internal.R
\name{meankNN}
\alias{meankNN}
\title{Mean kNN for one new point}
\usage{
meankNN(new, set.ref, k_all, ...)
}
\arguments{
\item{new}{The vector that represents the new point}

\item{set.ref}{The ensemble of points in the reference data set  (row by row : \code{set.ref[j,]} represents the j_th point)}

\item{k_all}{specify the values of k}

\item{...}{further arguments to be passed to the function (ignored).}
}
\value{
vector of mean kNN values for the new point with the respect to the reference data set, for each k in k_all.
}
\description{
Compute the mean kNN for the new point with respect to the data set of reference for various values of k.
}
\keyword{internal}
