% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_stats}
\alias{norm_stats}
\title{Normalize the reference table}
\usage{
norm_stats(target, sumstat, norm = sd)
}
\arguments{
\item{target}{a matrix of observations, with dimension Nobs x Nstats.}

\item{sumstat}{a matrix of summary statistics from the training set,
with dimension Ntrain x Nstats.}

\item{norm}{the normalization function. Default to \code{\link{sd}}.}
}
\value{
A list, with the normalized \code{target} and \code{sumstat} matrices.
}
\description{
This function applies the \code{norm} function to each of the Ntrain columns
of the training \code{sumstat} matrix. It then normalize each column of both the
\code{target} and \code{sumstat} matrices by dividing them by the norm:

\code{target[, j] = target[, j] / norm(sumstat[, j])} and
\code{sumstat[, j] = sumstat[, j] / norm(sumstat[, j])}
}
