% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgfit.R
\name{postgfit}
\alias{postgfit}
\title{Posterior goodness of fit function}
\usage{
postgfit(
  target,
  sumstat,
  sumstat.replica,
  nb.replicate = 100,
  score = c("lof", "kNN"),
  k = c(2, 3, 5, 8, 14, 20),
  k_range = range(k),
  eps = 0.01,
  norm = sd,
  ncores = 1
)
}
\arguments{
\item{target}{a matrix of observations, with dimension Nobs x Nstats.}

\item{sumstat}{a matrix of summary statistics from the training set,
with dimension Ntrain x Nstats.}

\item{sumstat.replica}{a matrix of replicate summary statistics from the training set,
with dimension Ntrain x Nstats.
Each line i of the replicate summary statistic should have been simulated
with the same parameters as the one used for the same line i of the reference table.}

\item{nb.replicate}{number of replicates to draw from the training set for the calibration of the p value.}

\item{score}{the score to use for calling outliers. Can be "lof", "kNN", or both.}

\item{k}{The kth-distance to be used to calculate LOFs of kNNs.
k can be a vector which contains multiple k values based on which the score needs to be calculated.}

\item{k_range}{A vector of min and max values for k when using the "LOF max" score (see Details).}

\item{eps}{the proportion of data points used to approximate the posterior distribution
of the reference table around each observation point. Default to 0.01.}

\item{norm}{the normalization function. Default to \code{\link{sd}}.}

\item{ncores}{number of cores for parallel computations.}
}
\value{
For each score, it a list with:
\itemize{
\item \code{score.calib} a matrix of LOF scores for the calibration data (rows), for each possible value of k.
\item \code{score.test} a matrix of LOF scores for the test data (rows), for each possible value of k.
\item \code{pval} a matrix of p-values for the test data (rows), for each possible value of k.
\item \code{n.ref} the number of particles in the reference dataset (number of rows of \code{sumstat}).
\item \code{n.calib} the number of particles used for calibration (equal to \code{nb.replicate}).
\item \code{k_range} the min and max values for k when using the "LOF max" score.
}
}
\description{
This function is likely to be DEPRECATED in stable versions of the package.

This function performs a prior Goodness of Fit (GoF) test,
that checks whether we can reject the hypothesis that the \code{target}
is from the same model as the particles from the \code{sumstat} matrix.

It works by sub-sampling \code{nb.replicate} points from a localized subset of the
reference table \code{sumstat}, and then by comparing the score of the
\code{target} with the distribution of scores of the calibration points.
}
\references{
Le Mailloux G., Bastide, P., Marin, J-M., Estoup, A. (2024+),
Goodness of Fit for Bayesian Generative Models with Applications in Population Genetics.

Markus M. Breunig, Hans-Peter Kriegel, Raymond T. Ng, and Jörg Sander. 2000.
LOF: identifying density-based local outliers. SIGMOD Rec. 29, 2 (June 2000), 93–104.
https://doi.org/10.1145/335191.335388
}
\seealso{
\code{\link[=gfit]{gfit()}} for pre-inference prior GoF test.
}
